+++ Mods ChoGGi/Dome Show Connected Objects/metadata.lua
@@ -12,7 +12,7 @@ return PlaceObj("ModDef", {
 	"version", 4,
 	"version_major", 0,
 	"version_minor", 4,
-	"image", "Preview.png",
+	"image", "Preview.jpg",
 	"steam_id", "2038895989",
 	"pops_any_uuid", "42e8c2ef-c183-4488-bce1-0b8b0223c7ba",
 	"author", "ChoGGi",
@@ -25,9 +25,9 @@ return PlaceObj("ModDef", {
 	"description", [[Wondering why you can't remove a dome? This mod will show any objects that will block you from doing so.
 
 Mod Options:
-Enable Mod: If you want to disable the mod without the mod manager.
-Move Invalid Position: Move any objects at an invalid position to the dome when you press the toggle connected objs button (default disabled).
-Clean Up Invalid: Remove any invalid objects stuck in the dome when you press the toggle connected objs button (default disabled).
+Enable Mod
+[b]Move Invalid Position[/b]: Move any objects at an invalid position to the dome when you press the toggle connected objs button (default disabled).
+[b]Clean Up Invalid Objects[/b]: Remove any invalid objects stuck in the dome when you press the toggle connected objs button (default disabled).
 
 If you see a white line going off to nothing, it's an invalid object (probably).
 ]],