+++ Mods ChoGGi/Find Map Locations/metadata.lua
@@ -18,9 +18,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2453011286",
 	"pops_any_uuid", "b97dbd4e-71cf-4e82-80a0-46dd104133bc",
 	"lua_revision", 1001569,
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {