+++ Mods ChoGGi/Fix Stuck Dredger Marks/metadata.lua
@@ -1,0 +1,20 @@
+return PlaceObj("ModDef", {
+	"title", "Fix Stuck Dredger Marks",
+	"id", "ChoGGi_FixStuckDredgerMarks",
+	"steam_id", "2453113211",
+	"pops_any_uuid", "8817cb9d-0795-469f-970c-37c7e257973f",
+	"lua_revision", 1001569,
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagOther", true,
+	"description", [[
+https://www.reddit.com/r/SurvivingMars/comments/mo3kjv/what_are_those_blue_circles/
+]],
+})